add kit1
